+++ packages/pluggableWidgets/notifications-native/CHANGELOG.md
@@ -10,6 +10,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 -   We fixed an issue that occured on Android devices when using the Make it Native app to preview a page containing the Notifications widget.
 
+### Added
+
+- Structure mode preview
+
 ## [3.1.1] - 2022-3-9
 
 ### Fixed